+++ example/lib/main.dart
@@ -4,21 +4,37 @@ import 'package:bybit/bybit.dart';
 Future<void> readWebSocket(Stream<dynamic> stream, int count) async {
   var i = 0;
   await for (var value in stream) {
-    print(value);
     i++;
+    print(value)
     if (i >= count) return;
   }
 }
 
 void main() async {
-  var bybit = ByBit(logLevel: 'DEBUG');
-  bybit.getSymbolsInfoPeriodic(period: Duration(seconds: 1));
+  var bybit = ByBit(logLevel: 'INFO');
+
+  // Define REST API calls that we want to make periodically
+  bybit.getServerTimePeriodic(period: Duration(seconds: 5));
+  bybit.getAnnouncementPeriodic(period: Duration(seconds: 5));
+  bybit.getOpenInterestPeriodic(
+      symbol: 'ETHUSD',
+      interval: '15min',
+      period: Duration(seconds: 2),
+      limit: 3);
+
+  // Connect to the Server
   bybit.connect();
-  //bybit.subscribeToKlines(symbol: 'BTCUSD', interval: 'D');
-  await readWebSocket(bybit.stream, 30);
-  //var symbols = await bybit.getSymbolsInfo();
-  //print(symbols);
-  var test = await bybit.getLiquidatedOrders(symbol: 'ETHUSD', limit: 3);
-  print(test);
+
+  // Subscribe to WebSockets channels
+  bybit.subscribeToKlines(symbol: 'BTCUSD', interval: 'D');
+
+  // Show the stream output
+  await readWebSocket(bybit.stream, 10);
+
+  // Once the 10 first server response are shown, make a single REST API call
+  var symbols = await bybit.getSymbolsInfo();
+  print(symbols);
+  
+  // Close sockets
   bybit.disconnect();
 }